Move common service functions to their corresponding files
