+++ src/modules/history/pages/HomePage.tsx
@@ -4,12 +4,15 @@ import { useEffect, useState } from 'react';
 import { useHistory } from 'react-router-dom';
 import { BrowserStorage } from '../../../common/BrowserStorage';
 import { I18N } from '../../../common/I18N';
-import { getServicePages, StreamingServicePage } from '../../../streaming-services/common/common';
+import {
+	StreamingService,
+	streamingServices,
+} from '../../../streaming-services/streaming-services';
 import { HistoryInfo } from '../components/HistoryInfo';
 
 export const HomePage: React.FC = () => {
 	const history = useHistory();
-	const [services, setServices] = useState([] as StreamingServicePage[]);
+	const [services, setServices] = useState([] as StreamingService[]);
 
 	const onRouteClick = (path: string) => {
 		history.push(path);
@@ -18,13 +21,9 @@ export const HomePage: React.FC = () => {
 	useEffect(() => {
 		const checkEnabledServices = () => {
 			const serviceOptions = BrowserStorage.options.streamingServices;
-			const enabledServices = [];
-			const servicePages = getServicePages();
-			for (const service of servicePages) {
-				if (service.hasSync && serviceOptions[service.id].sync) {
-					enabledServices.push(service);
-				}
-			}
+			const enabledServices = Object.values(streamingServices).filter(
+				(service) => service.hasSync && serviceOptions[service.id].sync
+			);
 			setServices(enabledServices);
 		};
 
@@ -42,7 +41,7 @@ export const HomePage: React.FC = () => {
 								key={service.id}
 								button={true}
 								divider={true}
-								onClick={() => onRouteClick(service.path)}
+								onClick={() => onRouteClick(`/${service.id}`)}
 							>
 								<ListItemText primary={service.name} />
 							</ListItem>
